rename also file
